+++ services/GateWay/src/core/domain/dto/user/user_dto.hpp
@@ -36,6 +36,32 @@ struct CORE_EXPORT UserSendCodeDTO
   }
 };
 
+struct CORE_EXPORT UserRegisterDTO
+{
+  std::string nickname;
+  std::string email;
+  std::string password;
+  std::string confirm_password;
+  std::string verify_code;
+  std::int8_t purpose;
+
+  static std::optional<UserRegisterDTO> FromJsonString(const std::string& json_str)
+  {
+    auto root = utils::ParseJson(json_str);
+    if (!root || !(*root).isObject())
+    {
+      return std::nullopt;
+    }
+
+    return UserRegisterDTO{.nickname = (*root)["nickname"].asString(),
+                           .email = (*root)["email"].asString(),
+                           .password = (*root)["password"].asString(),
+                           .confirm_password = (*root)["confirm_password"].asString(),
+                           .verify_code = (*root)["verify_code"].asString(),
+                           .purpose = static_cast<std::int8_t>((*root)["purpose"].asInt())};
+  }
+};
+
 }  // namespace core
 
 #endif  // USER_DTO_HPP