+++ services/GateWay/src/core/service/user/user_service.hpp
@@ -25,6 +25,7 @@ public:
   static UserService& GetInstance();
 
   void HandleSendCodeRequest(const utils::Context& ctx, const UserSendCodeDTO& dto, core::CommonVO& common_vo) const;
+  void HandleRegisterRequest(const utils::Context& ctx, const UserRegisterDTO& dto, core::CommonVO& common_vo) const;
 
   UserService(const UserService&) = delete;
   UserService& operator=(const UserService&) = delete;