+++ services/GateWay/src/utils/common/routes.hpp
@@ -21,10 +21,12 @@ namespace utils
 // ================
 constexpr const char* HEALTH_CHECK_ROUTE = "/health-check";
 constexpr const char* USER_SEND_CODE_ROUTE = "/user/send-code";
+constexpr const char* USER_REGISTER_ROUTE = "/user/register";
 
 inline const std::unordered_set<std::string> NO_AUTH_ROUTES = {
     HEALTH_CHECK_ROUTE,
     USER_SEND_CODE_ROUTE,
+    USER_REGISTER_ROUTE,
 };
 
 }  // namespace utils